# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface CheckWalletRequest {
  /** Ethereum address (0x + 40 hex) */
  address: string;
}

/**
 * Tier label derived from the score
 */
export type WalletStatsTier =
  (typeof WalletStatsTier)[keyof typeof WalletStatsTier];

export const WalletStatsTier = {
  unranked: "unranked",
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
  diamond: "diamond",
} as const;

export interface WalletStats {
  address: string;
  eligible: boolean;
  tx_count: number;
  contract_count: number;
  active_months: number;
  active_days: number;
  native_volume_eth: number;
  gasfee_eth: number;
  current_balance_eth: number;
  /** ISO timestamp of first transaction */
  first_activity?: string | null;
  /** ISO timestamp of latest transaction */
  last_activity?: string | null;
  /**
   * Composite 0-1000 score derived from activity metrics
   * @minimum 0
   * @maximum 1000
   */
  score: number;
  /** Tier label derived from the score */
  tier: WalletStatsTier;
}

export interface EligibilityCriteria {
  contract_count: number;
  tx_count: number;
  active_months: number;
  native_volume_eth: number;
}

export interface RecentCheck {
  address: string;
  score: number;
  tier: string;
  eligible: boolean;
  /** ISO timestamp */
  checked_at: string;
}

export type GetRecentChecksParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type GetRecentChecks200 = {
  items: RecentCheck[];
};