# lib/api-zod/src/generated/types/checkWalletRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckWalletRequest {
  /** Ethereum address (0x + 40 hex) */
  address: string;
}