# lib/api-zod/src/generated/types/eligibilityCriteria.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EligibilityCriteria {
  contract_count: number;
  tx_count: number;
  active_months: number;
  native_volume_eth: number;
}