# lib/api-zod/src/generated/types/recentCheck.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentCheck {
  address: string;
  score: number;
  tier: string;
  eligible: boolean;
  /** ISO timestamp */
  checked_at: string;
}