# lib/api-zod/src/generated/types/walletStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WalletStatsTier } from "./walletStatsTier";

export interface WalletStats {
  address: string;
  eligible: boolean;
  tx_count: number;
  contract_count: number;
  active_months: number;
  active_days: number;
  native_volume_eth: number;
  gasfee_eth: number;
  current_balance_eth: number;
  /** ISO timestamp of first transaction */
  first_activity?: string | null;
  /** ISO timestamp of latest transaction */
  last_activity?: string | null;
  /**
   * Composite 0-1000 score derived from activity metrics
   * @minimum 0
   * @maximum 1000
   */
  score: number;
  /** Tier label derived from the score */
  tier: WalletStatsTier;
}